factors: Reorganize spin-factors
