+++ crates/factor-outbound-networking/Cargo.toml
@@ -5,7 +5,6 @@ authors = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-anyhow = "1"
 futures-util = "0.3"
 ipnet = "2.9.0"
 spin-factor-variables = { path = "../factor-variables" }
