+++ crates/factor-wasi/Cargo.toml
@@ -5,7 +5,6 @@ authors = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-anyhow = "1"
 cap-primitives = "3.0.0"
 spin-factors = { path = "../factors" }
 wasmtime-wasi = { workspace = true }
