+++ crates/factor-wasi/src/preview1.rs
@@ -1,5 +1,6 @@
 use spin_factors::{
-    Factor, FactorInstancePreparer, InitContext, InstancePreparers, PrepareContext, Result, SpinFactors
+    anyhow, Factor, FactorInstancePreparer, InitContext, InstancePreparers, PrepareContext,
+    SpinFactors,
 };
 use wasmtime_wasi::{preview1::WasiP1Ctx, WasiCtxBuilder};
 
@@ -10,7 +11,10 @@ impl Factor for WasiPreview1Factor {
     type InstancePreparer = InstancePreparer;
     type InstanceState = WasiP1Ctx;
 
-    fn init<Factors: SpinFactors>(&mut self, mut ctx: InitContext<Factors, Self>) -> Result<()> {
+    fn init<Factors: SpinFactors>(
+        &mut self,
+        mut ctx: InitContext<Factors, Self>,
+    ) -> anyhow::Result<()> {
         ctx.link_module_bindings(wasmtime_wasi::preview1::add_to_linker_async)
     }
 }
@@ -23,13 +27,13 @@ impl FactorInstancePreparer<WasiPreview1Factor> for InstancePreparer {
     fn new<Factors: SpinFactors>(
         _ctx: PrepareContext<WasiPreview1Factor>,
         _preparers: InstancePreparers<Factors>,
-    ) -> Result<Self> {
+    ) -> anyhow::Result<Self> {
         Ok(Self {
             wasi_ctx: WasiCtxBuilder::new(),
         })
     }
 
-    fn prepare(mut self) -> Result<<WasiPreview1Factor as Factor>::InstanceState> {
+    fn prepare(mut self) -> anyhow::Result<WasiP1Ctx> {
         Ok(self.wasi_ctx.build_p1())
     }
 }
